Moving mixup and cutmix to references.
